Refactor 9: From Switch/Case to Polymorphism This one was by far the largest and most drastic refactoring, yet yields amazing results!  What we did here was take the price aspect out of Movie and created an abstract price class.  Then we created several derived types of Price (Regular, Childrens, and NewRelease)  This allows us to add new types anytime we want and there are simply no strings attached.  Since the movie contains a price object, that object contains all the information it needs to give the movie class what it wants, no links are broken when we add/remove a movie type.
The reason for making price an abstract class, rather than an interface, is in the case of NewRelease, as it is the only movie type (currently) that yields a different number of renter bonus days.  The others simply call the generic method found in Price.

While performing this refactoring, the test cases came in much more useful than before, as they failed! I was amazed and quite curious about how this could happen.  Upon further investigation, I found the following things:
A. I was missing [TestMethod] in front of 2 of my test methods, causing them not to be evaluated.
B. Part A led me to find an incorrect use of terms in the translation from Java to C#.  I had simply specified "new" on NewReleasePrice's GetFrequentRenterPoints method when what I meant to do was make Price's GetFrequentRenterPoints virtual and then perform an override on it.  This was a very simple mistake that I would NOT have found (the code compiled and ran fine but would always return 1 for renter points) if I had not written the proper tests!